Fix ThreadAnalyzer build errors: Remove unavailable ClrMD API calls
Removed attempts to query ClrThreadPool telemetry (QueueLength, ActiveWorkerThreads,
IdleWorkerThreads, MinWorkerThreads, MaxWorkerThreads) and ClrThread.Name, which
are not available in ClrMD 4.

These properties are not exposed by the current Microsoft.Diagnostics.Runtime API.
ThreadPool telemetry would require direct memory inspection of runtime internals.
Thread naming would require enumerating managed thread objects.

Corresponding domain result properties are initialized to default values (0 for
counts). This allows P1-1 (ThreadPool metrics) and P1-2 (thread names) to be
shelved as "awaiting ClrMD API expansion" rather than implemented with stubs.

Also fixed ExecutiveSummaryProjector.cs type conversion error (long to int cast).

Note: Pre-existing build errors remain in ReferenceChainOptions (MaxPathSearchObjects,
FastModeMaxDepth properties not found) - unrelated to this work.

+++ src/DumpDetective.Reporting/Services/ExecutiveSummaryProjector.cs
@@ -189,8 +189,8 @@ internal sealed class ExecutiveSummaryProjector
     {
         for (int i = 0; i < runs.Count; i++)
         {
-            if (runs[i].Result is MemoryDomainResult mem) return mem.TotalObjects;
-            if (runs[i].Result is GCGenerationDomainResult gc) return gc.TotalObjects;
+            if (runs[i].Result is MemoryDomainResult mem) return (int)mem.TotalObjects;
+            if (runs[i].Result is GCGenerationDomainResult gc) return (int)gc.TotalObjects;
         }
 
         return null;
